+++ controllers/addContact.js
@@ -1,0 +1,8 @@
+const { Contact } = require("../models/contact");
+
+const createContact = async (req, res) => {
+  const result = await Contact.create(req.body);
+  res.status(201).json(result);
+};
+
+module.exports = createContact;
