--- controllers/contacts.js
@@ -1,64 +1,0 @@
-const { Contact } = require("../models/contact");
-const { HttpError, ctrlWrapper } = require("../helpers");
-
-const getAll = async (req, res) => {
-  const result = await Contact.find();
-  res.status(200).json(result);
-};
-
-const getById = async (req, res) => {
-  const { contactId } = req.params;
-  console.log("id", contactId);
-  const result = await Contact.findById(contactId);
-
-  if (!result) {
-    throw HttpError(404, `${req.params.contactId} not found`);
-  }
-
-  res.status(200).json(result);
-};
-
-const removeContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
-  if (!result) {
-    throw HttpError(404, `${req.params.contactId} not found`);
-  }
-  res.status(200).json({ message: "contact deleted" });
-};
-
-const createContact = async (req, res) => {
-  const result = await Contact.create(req.body);
-  res.status(201).json(result);
-};
-
-const updateContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, `${req.params.contactId} not found`);
-  }
-  res.status(200).json(result);
-};
-
-const updateStatusContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, `${req.params.contactId} not found`);
-  }
-  res.status(200).json(result);
-};
-
-module.exports = {
-  getAll: ctrlWrapper(getAll),
-  getById: ctrlWrapper(getById),
-  removeContact: ctrlWrapper(removeContact),
-  createContact: ctrlWrapper(createContact),
-  updateContact: ctrlWrapper(updateContact),
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-};
