+++ routes/api/contacts.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const ctrl = require("../../controllers/contacts");
+const ctrl = require("../../controllers");
 const { validateBody, isValidId } = require("../../middlewares");
 const { schemas } = require("../../models/contact");
 const validateBodyStatusUpdate = require("../../middlewares/validateBodyStatusUpdate");
